Cleaning crt and key temp files only in dev

+++ scripts/test-cf-openvidu-all-3.sh
@@ -338,5 +338,8 @@ fi
 # Cleaning
 rm $TEMPFILE
 rm $TEMPJSON
-rm $TEMPCRT
-rm $TEMPKEY
+
+if [ "${MODE}" == "dev" ]; then
+  rm $TEMPCRT
+  rm $TEMPKEY
+fi
